workers now return the path with their data.
before, a worker handler returned a dict. then a separate
function was called to figure out where in the document
that data dict should be merged.

We have simplified this and made it much more generalized.
now a worker handler returns a hash that contains the data
hash and the path at which to merge that data hash.